backend/requirements.txt
Django==4.0.5
django-cors-headers==3.13.0
uwsgi==2.0.17.1
django-nextjs==2.2.1
djangorestframework==3.14.0
django-nextjs==2.2.1
django-cors-headers==3.13.0
PyJWT==1.7.1
dj-rest-auth==2.2.5
django-allauth==0.51.0
Pillow==9.0.1